+++ SvnDockAppTests/OperationRecordRegressionTests.swift
@@ -1,0 +1,7 @@
+import XCTest
+
+final class OperationRecordRegressionTests: XCTestCase {
+    func testOperationRecordOutcomesCopyingAndRetention() throws {
+        try OperationRecordRegressionChecks.run()
+    }
+}